LPS-167653 Move children elements to Listing
